// lbo-sdk/src/plugin/cms/models/LayoutGameEntryTypeEnum.ts
export enum LayoutWidgetEntryTypeEnum {
    BANNER = "banner",
    GRID = "grid",
    TILES = "tiles",
    TOP_GAMES = "topGames",
    ATOZ = "atoz",
    RECENTLY_PLAYED_GAMES = "recentlyPlayedGames",
    RECOMMENDED_GAMES = "recommendedGames",
    TAGGED_GAMES = "taggedGames",
    JACKPOT_GRID = "jackpotGrid",
    JACKPOT_TILE = "jackpotTile",
    DFG = "dfg"
}